+++ mink/core/resource.py
@@ -3,7 +3,7 @@
 from enum import Enum
 from typing import Any
 
-from mink.core import exceptions
+from mink.core import exceptions, utils
 from mink.sparv import storage
 
 
@@ -32,6 +32,7 @@ class Resource:
         name: dict | None = None,
         type: ResourceType | str = ResourceType.corpus,  # noqa: A002
         source_files: list | None = None,
+        sources_deleted: str = "",
     ) -> None:
         """Init resource by setting class variables.
 
@@ -41,6 +42,7 @@ class Resource:
             name: The name of the resource.
             type: The type of the resource.
             source_files: List of source files.
+            sources_deleted: Timestamp of when sources were last deleted (used for knowing what to re-annotate).
         """
         self.id = id
         self.public_id = public_id or self.id
@@ -55,6 +57,7 @@ class Resource:
         else:
             raise exceptions.InvalidResourceTypeError(type)
         self.source_files = source_files or []
+        self.sources_deleted = sources_deleted or ""
 
     def __str__(self) -> str:
         """Return a string representation of the object by serializing it.
@@ -76,6 +79,7 @@ class Resource:
             "name": self.name,
             "type": self.type,
             "source_files": self.source_files,
+            "sources_deleted": self.sources_deleted,
         }
 
     def set_parent(self, parent: Any) -> None:
@@ -95,7 +99,13 @@ class Resource:
         self.name = name
         self.parent.update()
 
-    def set_source_files(self) -> None:
-        """Set source files and save."""
+    def set_source_files(self, deleted_sources: bool = False) -> None:
+        """Set 'source_files' list (and 'sources_deleted' timestamp) and save.
+
+        Args:
+            deleted_sources: Whether source files have been deleted.
+        """
         self.source_files = storage.list_contents(storage.get_source_dir(self.id))
+        if deleted_sources:
+            self.sources_deleted = utils.get_current_time()
         self.parent.update()